refactor(simd): deduplicate scalar fallbacks and tighten docs/asserts
